Delete ex folders
