--- Hit-Run-Reborn/Controller/.gitignore
@@ -1,21 +1,0 @@
-.vscode/*
-
-.vscode/.browse.c_cpp.db*
-.vscode/c_cpp_properties.json
-.vscode/launch.json
-
-
-.vscode/settings.json
-.vscode/tasks.json
-.vscode/launch.json
-.vscode/extensions.json
-
-
-.pio
-.pioenvs
-.piolibdeps
-
-.travis.yml
-
-.clang_complete
-.gcc-flags.json
