+++ WeaponSender/lib/README
