+++ WeaponSender/platformio.ini
@@ -8,12 +8,7 @@
 ; Please visit documentation for the other options and examples
 ; https://docs.platformio.org/page/projectconf.html
 
-; [env:nanoatmega328]
-; platform = atmelavr
-; board = nanoatmega328
-; framework = arduino
-
-[env:uno]
+[env:nanoatmega328]
 platform = atmelavr
-board = uno
+board = nanoatmega328
 framework = arduino
