+++ WeaponSender/src/main.cpp
@@ -1,69 +1,67 @@
+#include <Arduino.h>
+
 #include <SPI.h>
+
 #include "nRF24L01.h"
 #include "RF24.h"
 #include "printf.h"
 
+#define PIN_NRF_CE  7   // 7 without adapter , 9
+#define PIN_NRF_CSN 8  // 8 without adapter , 10
 
-#define PIN_BUTTON 4
-#define PIN_LED 2
+#define PIN_LED 4
+#define PIN_BUTTON 2
 
-// WITH ADAPTER
-// #define CE_PIN  9
-// #define CSN_PIN 10
-// END WITH
-
-// WITHOUT ADAPTER
-#define CE_PIN  7
-#define CSN_PIN 8
+/** Sportsmen Colors
+ *  1 - GREEN
+ *  2 - RED
+ */
+const byte SPORTSMEN_COLOR = 1;
 
 
 // Initialize
-RF24 radio(CE_PIN, CSN_PIN);
-byte counter = 0;
+RF24 radio(PIN_NRF_CE, PIN_NRF_CSN);
+
 byte buttonState = 0;
-// END
-
-void setup()
-{
-  // <--- SERIAL SETTINGS --->
-  Serial.begin(9600);
-  printf_begin();
-  printf("\n\r Address: 0xAABBCCDD22LL, sportsmen - GREEN \n\r");
-  printf("\n\r <----- EPEE SENDER -----> \n\r");
+byte blankMessage = 9;
 
 
-  // <--- RADIO SETTINGS --->
-  radio.begin();
+// Setup
+void setup() {
 
-  radio.setChannel(5);                                       // Указываем канал передачи данных (от 0 до 127), 5 - значит передача данных осуществляется на частоте 2,405 ГГц (на одном канале может быть только 1 приёмник и до 6 передатчиков)
-  radio.setDataRate     (RF24_1MBPS);                        // Указываем скорость передачи данных (RF24_250KBPS, RF24_1MBPS, RF24_2MBPS), RF24_1MBPS - 1Мбит/сек
-  radio.setPALevel      (RF24_PA_HIGH);                      // Указываем мощность передатчика (RF24_PA_MIN=-18dBm, RF24_PA_LOW=-12dBm, RF24_PA_HIGH=-6dBm, RF24_PA_MAX=0dBm)
-  radio.openWritingPipe (0xAABBCCDD11LL);                    // NANO
-  // radio.openWritingPipe (0xAABBCCDD22LL);                    // UNO
+        // Serial settings
+        Serial.begin(9600);
+        printf_begin();
+        printf("\n\r <<--- SENDER --->> \n\r");
 
-  // radio.stopListening();
+        // Radio settings
+        radio.begin();
+        radio.setChannel(5);                      // Указываем канал передачи данных (от 0 до 127), 5 - значит передача данных осуществляется на частоте 2,405 ГГц (на одном канале может быть только 1 приёмник и до 6 передатчиков)
+        radio.setDataRate(RF24_1MBPS);            // Указываем скорость передачи данных (RF24_250KBPS, RF24_1MBPS, RF24_2MBPS), RF24_1MBPS - 1Мбит/сек
+        radio.setPALevel(RF24_PA_HIGH);           // Указываем мощность передатчика (RF24_PA_MIN=-18dBm, RF24_PA_LOW=-12dBm, RF24_PA_HIGH=-6dBm, RF24_PA_MAX=0dBm)
+        radio.openWritingPipe(0xAABBCCDD11LL);    // NANO
+        // radio.openWritingPipe (0xAABBCCDD22LL);   // UNO
 
-  // BUTTON SETTINGS
-  pinMode(PIN_BUTTON, INPUT_PULLUP);
-
-  // LED SETTINGS
-  pinMode(PIN_LED, OUTPUT);
+        // Wait for console opening
+        delay(3000);
 
 }
 
-void loop(void)
-{
 
-  /********************** SENDER *******************************/
-  buttonState = digitalRead(PIN_BUTTON);
-  if (buttonState == LOW)
-  {
-    Serial.print("Button\n");
-    counter++;
-    radio.write(&counter, sizeof(counter));
-    printf("%d\n", counter);
-    digitalWrite(PIN_LED, HIGH);
-    delay(2000);
-    digitalWrite(PIN_LED, LOW);
-  }
+// Loop
+void loop() {
+
+    /*
+    Для шпаги, при нажатии кнопки, сигнал проходит до Sender'a
+    */
+    buttonState = digitalRead(PIN_BUTTON);
+
+    if (buttonState == HIGH)
+    {
+        radio.write(&SPORTSMEN_COLOR, sizeof(SPORTSMEN_COLOR));
+        printf("Button clicked!\n");
+        digitalWrite(PIN_LED, HIGH);
+        delay(2000);
+        digitalWrite(PIN_LED, LOW);
+    }
 }
